backend/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
flask-restful
flask_sqlalchemy
flask_migrate
flask_jwt_extended
